backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.0
sqlmodel==0.0.19
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
pydantic-settings==2.3.0
alembic==1.13.1
